change integer to long

+++ src/main/java/com/google/graphgeckos/dashboard/datatypes/BuildBotData.java
@@ -17,7 +17,6 @@ package com.google.graphgeckos.dashboard.datatypes;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.google.cloud.Timestamp;
 import java.util.ArrayList;
-import java.util.Date;
 import java.util.List;
 import java.util.Map;
 import org.springframework.cloud.gcp.data.datastore.core.mapping.Entity;
@@ -91,7 +90,7 @@ public class BuildBotData {
   @JsonProperty("sourceStamp")
   public void unpackSourceStamp(Map<String, Object> sourceStamp) {
     commitHash = sourceStamp.get("revision").toString();
-    timestamp = Timestamp.ofTimeMicroseconds(Integer.parseInt(sourceStamp.get("when").toString()));
+    timestamp = Timestamp.ofTimeMicroseconds(Long.parseLong(sourceStamp.get("when").toString()));
   }
 
   /**
